Add STS support for temporary s3 sessions
STS requires a session token to be provided along with the key and
secret. This change allows the user to provide the session token in both
versions of the log4j plugin. The value defaults to null to preserve the
original API.

Testing was performed manually by using this plugin to push Spark logs
to a s3 bucket.